refactor: removed some unnecessary code repetition

+++ CapWatchBackend.Application/Exceptions/StoreTypeInvalidException.cs
@@ -6,7 +6,7 @@ namespace CapWatchBackend.Application.Exceptions {
   [Serializable]
   public class StoreTypeInvalidException : BaseException {
     public StoreTypeInvalidException() : base("Der Typ ist ungültig.") {
-      Status = (int)HttpStatusCode.BadRequest;
+      Status = HttpStatusCode.BadRequest;
     }
 
     protected StoreTypeInvalidException(SerializationInfo info, StreamingContext context) : base(info, context) { }
